refactor: move irf rows to separate file
